% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_MLE_hetero.R
\name{lin_MLE_hetero}
\alias{lin_MLE_hetero}
\title{Lin and Stiver's MLE Test with Heteroscedasticity}
\usage{
lin_MLE_hetero(X, y = NULL, alt = c("two.sided", "less", "greater"))
}
\arguments{
\item{X}{A 2 x k (or k x 2) matrix with partially matched pairs. If parameter y is specified, then X should instead be a vector with length equal to y.}

\item{y}{A vector of data paired with the X vector of data. Length should be equal to length of X. Default is NULL if X is a matrix of the correct dimensions.}

\item{alt}{The test's alternative hypothesis. Deaults to 'two.sided'.}
}
\value{
The test statistic (z.stat) and the p-value (p.val) of the test.
}
\description{
Performs an MLE based test for heteroscedastic partially matched samples with a matrix or two-vector input that includes missing data.
}
\examples{
#INSTALL THE LIBRARY 'mvtnorm'
#Generate bivariate gaussian sample data

require(mvtnorm)
set.seed(123) # simulation seed

rho <- c(0.2, 0.7) # correlation test values
n <- c(20, 50) # test sample sizes
sig1 <- c(1, 3) # heteroscedastic test standard deviations


# Heterosedastic data with n=50, rho=0.7

hetero_50_0.7 <- rmvnorm(n[2], mean=c(0, 0), sigma=matrix(c(sig1[1]^2, rho[2]*sig1[1]*sig1[2], rho[2]*sig1[1]*sig1[2], sig1[2]^2), nrow=2, byrow=T))


#Generate indices to make NA
NA_index50 <- sample(1:50, 20) # 20 NA indices generated


#Use generated indices to make random entries NA

hetero_50_0.7[,1][NA_index50[1:10]] <- NA
hetero_50_0.7[,2][NA_index50[11:20]] <- NA


#We now have a sample data set called hetero_50_0.7 that is heteroscedastic, n=50, and rho=0.7. A complete dataset for testing can be found in the vignette.


#Example of function use, two vectors

lin_MLE_hetero(hetero_50_0.7[,1], hetero_50_0.7[,2], alt='less')

#Example of function use, one matrix

lin_MLE_hetero(hetero_50_0.7, alt='less')

}
